Provide an example for an HMI service
HMIs will be provided to the platform as services that serve a frontend over HTTP. This is a simple HMI written in Golang. It serves a frontend that has some executive-control functionality.

A guide that references this example code will be added to the developer documentation on https://flowstate.intrinsic.ai/docs/. This code uses that guide as a base and adds a lot of additional functionality